vizoci/pyproject.toml
[project]
name = "vizoci"
version = "0.1.0"
description = "Oracle Cloud Infrastructure Instance Creator"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.0",
    "oci>=2.0.0",
    "httpx>=0.25.0",
    "click>=8.1.0",
    "rich>=13.0.0",
]

[project.scripts]
vizoci = "vizoci.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"